+++ src/tcbs/listing.ts
@@ -5,4 +5,20 @@ export default class Listing {
     // TODO: Implement this method
     return Promise.reject(new Error("Currently this not implemented"));
   }
+  async symbolsByExchange() {
+    // TODO: Implement this method
+    return Promise.reject(new Error("Currently this not implemented"));
+  }
+  async symbolsByIndustries() {
+    // TODO: Implement this method
+    return Promise.reject(new Error("Currently this not implemented"));
+  }
+  async industriesIcb() {
+    // TODO: Implement this method
+    return Promise.reject(new Error("Currently this not implemented"));
+  }
+  async symbolsByGroup({ group = "VN30" }: { group?: string }) {
+    // TODO: Implement this method
+    return Promise.reject(new Error("Currently this not implemented"));
+  }
 }