+++ src/vci/model.ts
@@ -105,3 +105,55 @@ export interface ITicker {
     organ_name: string;
   };
 }
+
+export interface ITickerInfo {
+  ticker: string;
+  organName: string;
+  enOrganName: string;
+  icbName3: string;
+  enIcbName3: string;
+  icbName2: string;
+  enIcbName2: string;
+  icbName4: string;
+  enIcbName4: string;
+  comTypeCode: string;
+  icbCode1: string;
+  icbCode2: string;
+  icbCode3: string;
+  icbCode4: string;
+  __typename: string;
+}
+
+export interface IFinancialRatio {
+  id: string;
+  type: string; // *
+  name: string; // *
+  unit: string; // *
+  isDefault: boolean;
+  fieldName: string;
+  en_Type: string;
+  en_Name: string; // *
+  tagName: string;
+  comTypeCode: string; // *
+  order: string; // *
+  __typename: string;
+}
+
+export interface IIcb {
+  icbCode: string;
+  level: string;
+  icbName: string;
+  enIcbName: string;
+  __typename: string;
+}
+
+export interface ISymbol {
+  id: number;
+  symbol: string;
+  type: string;
+  board: string;
+  enOrganName: string;
+  enOrganShortName: string;
+  organShortName: string;
+  organName: string;
+}