+++ src/vnstock.test.ts
@@ -15,9 +15,51 @@ describe("VCI", () => {
     expect(result).toHaveLength(2);
   });
 
-  test("VCI Listing all symbols", async () => {
+  test("VCI Listing allSymbols", async () => {
     const { stock } = new Vnstock();
     const result = await stock.listing.allSymbols();
+    expect(result).toBeDefined();
+  });
+
+  test("VCI Listing symbolsByExchange", async () => {
+    const { stock } = new Vnstock();
+    const result = await stock.listing.symbolsByExchange();
+    expect(result).toBeDefined();
+  });
+
+  test("VCI Listing symbolsByIndustries", async () => {
+    const { stock } = new Vnstock();
+    const result = await stock.listing.symbolsByIndustries();
+    expect(result).toBeDefined();
+  });
+
+  test("VCI Listing industriesIcb", async () => {
+    const { stock } = new Vnstock();
+    const result = await stock.listing.industriesIcb();
+    expect(result).toBeDefined();
+  });
+
+  test("VCI Listing symbolsByGroup", async () => {
+    const { stock } = new Vnstock();
+    const result = await stock.listing.symbolsByGroup({});
+    expect(result).toBeDefined();
+  });
+
+  test("VCI Financials balanceSheet", async () => {
+    const { stock } = new Vnstock();
+    const result = await stock.financials.balanceSheet({ symbol: symbols[0] });
+    expect(result).toBeDefined();
+  });
+
+  test("VCI Financials incomeStatement", async () => {
+    const { stock } = new Vnstock();
+    const result = await stock.financials.incomeStatement({ symbol: symbols[0] });
+    expect(result).toBeDefined();
+  });
+
+  test("VCI Financials cashFlow", async () => {
+    const { stock } = new Vnstock();
+    const result = await stock.financials.cashFlow({ symbol: symbols[0] });
     expect(result).toBeDefined();
   });
 });
@@ -40,10 +82,16 @@ describe("TCBS", () => {
   });
 });
 
-describe("Gold", () => {
+describe("Commodity", () => {
   test("Gold Price", async () => {
     const { commodity } = new Vnstock();
     const result = await commodity.goldPrice();
+    expect(result).not.toBeNull();
+  });
+
+  test("VCB Exchange Rates", async () => {
+    const { commodity } = new Vnstock();
+    const result = await commodity.exchangeRates();
     expect(result).not.toBeNull();
   });
 });