+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## vnstock-js 0.2.0 release
+Release with the following improvements:
+
+* Support **VCI** source
+  * Listing: **Symbols By Exchange**, **Symbols By Industries**, **Industry ICB** ,**Symbols By Group**.
+  * Financials Report: **Balance Sheet**, **Income Statement**, **Cashflow**.
+* Support **Exchange Rates** on Commodity Prices.
+
 ## vnstock-js 0.1.0 release
 
 Initial release with the following features: